add createProduct API
